Add verbose_name to trackers.

+++ android/src/com/firebirdberlin/tinytimetracker/SQLiteHandler.java
@@ -11,21 +11,23 @@ public class SQLiteHandler extends SQLiteOpenHelper {
   public static final String COLUMN_ID = "_id";
   public static final String COLUMN_METHOD = "method";
   public static final String COLUMN_NAME = "name";
+  public static final String COLUMN_VERBOSE = "verbose_name";
 
   public static final String TABLE_LOGS = "logs";
-  public static final String COLUMN_TRACKER_ID = "tracker_id";
-  public static final String COLUMN_TIMESTAMP_START = "timestamp_start";
   public static final String COLUMN_TIMESTAMP_END = "timestamp_end";
+  public static final String COLUMN_TIMESTAMP_START = "timestamp_start";
+  public static final String COLUMN_TRACKER_ID = "tracker_id";
 
   private static final String DATABASE_NAME = "trackers.db";
-  private static final int DATABASE_VERSION = 1;
+  private static final int DATABASE_VERSION = 2;
 
   // Database creation sql statement
   private static final String DATABASE_CREATE_TRACKERS =
         "CREATE TABLE " + TABLE_TRACKERS + "("
           + COLUMN_ID + " integer primary key autoincrement, "
-          + COLUMN_METHOD+ " text not null, "
-          + COLUMN_NAME + " text not null);";
+          + COLUMN_METHOD + " text not null, "
+          + COLUMN_NAME + " text not null, "
+          + COLUMN_VERBOSE + " text not null);";
   private static final String DATABASE_CREATE_LOGS =
         "CREATE TABLE " + TABLE_LOGS + "("
           + COLUMN_ID + " integer primary key autoincrement, "
@@ -48,6 +50,13 @@ public class SQLiteHandler extends SQLiteOpenHelper {
       Log.w(SQLiteHandler.class.getName(),
               "Upgrading database from version " + oldVersion + " to "
               + newVersion + ", which will destroy all old data");
+
+      if (oldVersion == 1 && newVersion == 2) {
+          db.execSQL("ALTER TABLE " + TABLE_TRACKERS + " ADD COLUMN " + COLUMN_VERBOSE + " TEXT DEFAULT '' NOT NULL");
+          db.execSQL("UPDATE " + TABLE_TRACKERS + " SET " + COLUMN_VERBOSE + " = " + COLUMN_NAME);
+          return;
+      }
+      // otherwise drop and re-create
       db.execSQL("DROP TABLE IF EXISTS " + TABLE_TRACKERS);
       db.execSQL("DROP TABLE IF EXISTS " + TABLE_LOGS);
       onCreate(db);
